move context to avoid hot-reload issue
